mapsforge-themes change package, rename InternalRenderTheme to MapsforgeThemes
